# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotel Booking Service",
    "description": "Микросервис для бронирования отелей: отели, комнаты, пользователи и бронирования",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api/hotel-booking-service"
    }
  ],
  "paths": {
...
    "/api/v1/rooms/": {
      "post": {
        "tags": ["Rooms"],
        "summary": "Create Room",
        "description": "Создание комнаты",
        "operationId": "create_room_api_v1_rooms__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_RoomRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Rooms"],
        "summary": "List Rooms",
        "description": "Реестр комнат",
        "operationId": "list_rooms_api_v1_rooms__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_List_RoomWithHotelRead__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rooms/{room_id}": {
      "get": {
        "tags": ["Rooms"],
        "summary": "Get Room",
        "operationId": "get_room_api_v1_rooms__room_id__get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_RoomWithHotelRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Rooms"],
        "summary": "Update Room",
        "operationId": "update_room_api_v1_rooms__room_id__put",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_RoomRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Rooms"],
        "summary": "Delete Room",
        "operationId": "delete_room_api_v1_rooms__room_id__delete",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Информация об ошибке"
      },
...
      "HotelRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "HotelRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RoomCreate": {
        "properties": {
          "hotel_id": {
            "type": "integer",
            "title": "Hotel Id"
          },
          "room_number": {
            "type": "string",
            "title": "Room Number"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_per_night": {
            "type": "number",
            "title": "Price Per Night"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "amenities": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              },
              {
                "type": "null"
              }
            ],
            "default": "available"
          }
        },
        "type": "object",
        "required": [
          "hotel_id",
          "room_number",
          "floor",
          "room_type",
          "price_per_night",
          "capacity"
        ],
        "title": "RoomCreate",
        "description": "Схема для создания комнаты"
      },
      "RoomRead": {
        "properties": {
          "room_number": {
            "type": "string",
            "title": "Room Number"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_per_night": {
            "type": "number",
            "title": "Price Per Night"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "amenities": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "hotel_id": {
            "type": "integer",
            "title": "Hotel Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "hotel_id",
          "created_at"
        ],
        "title": "RoomRead"
      },
      "RoomStatus": {
        "type": "string",
        "enum": [
          "available",
          "occupied",
          "maintenance",
          "cleaning"
        ],
        "title": "RoomStatus"
      },
      "RoomUpdate": {
        "properties": {
          "room_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Number"
          },
          "floor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "room_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_per_night": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Night"
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "amenities": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RoomUpdate"
      },
      "RoomWithHotelRead": {
        "properties": {
          "room_number": {
            "type": "string",
            "title": "Room Number"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_per_night": {
            "type": "number",
            "title": "Price Per Night"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "amenities": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amenities"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "hotel_id": {
            "type": "integer",
            "title": "Hotel Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "hotel": {
            "$ref": "#/components/schemas/HotelRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "hotel_id",
          "created_at"
        ],
        "title": "RoomWithHotelRead"
      },
...
      "SuccessResponse_List_RoomWithHotelRead__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/RoomWithHotelRead"
            },
            "type": "array",
            "title": "Data"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[List[RoomWithHotelRead]]"
      },
...
      "SuccessResponse_RoomRead_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/RoomRead"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[RoomRead]"
      },
      "SuccessResponse_RoomWithHotelRead_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/RoomWithHotelRead"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[RoomWithHotelRead]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}